requirements.txt
google-generativeai>=0.8.0
google-adk>=0.1.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
